Finish basic frame drawing functionality
I changed the frame number text box to a spin box. This is better
for the user and for the code.

I connected signals from the slider and spin box to slots in the
main windows. The two widgets make the sequence advance
or back up like a video display.

I added the ground truth bounding boxes to the sequence class.

I added functionality to draw the current frame's ground truth
bounding box on the frame image.