Refactor CMake test data setup
I added a CMake function, setup_test_data(), to consistently run
commands for setting up test data. The function checks the result
of the command, and prints the command output if the command
failed.

This is troubleshoot #20. That bug is difficult to reproduce. I'm
committing this to the main branch, then I'll watch GH workflows
during development of other features for this bug.