Refactor application logic
I moved application logic, such as the application settings and the
dataset, to an application class. This code was in the main window,
but it doesn't belong there. This cleans up the main window class.

I moved chart related code into two new files, also to clean up the
main window files. The new files are .hidden so clang-tidy doesn't
scan them.

closes #8