Draw tracking paths on the frame display
I added functionality to draw tracking paths on the
frame display. The user can select between tracking
paths and bounding boxes with a new combobox.

Tracking paths draws a polyline through the center
of each bounding box for the frames in the sequence.
The current point in the path is marked by a larg dot.
The future part of the path is dimmed with 1/4 opacity.

closes #37